conformance/buffimgs/cl_commandQueue_enqueueReadBufferRect.html

<!--
 Copyright (C) 2013 Samsung Electronics Corporation. All rights reserved.

 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided the following condition
 are met:

 1.  Redistributions of source code must retain the above copyright
     notice, this list of conditions and the following disclaimer.

 2.  Redistributions in binary form must reproduce the above copyright
     notice, this list of conditions and the following disclaimer in the
     documentation and/or other materials provided with the distribution.

 THIS SOFTWARE IS PROVIDED BY SAMSUNG ELECTRONICS CORPORATION AND ITS
 CONTRIBUTORS "AS IS", AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING
 BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
 FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL SAMSUNG
 ELECTRONICS CORPORATION OR ITS CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
 INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES(INCLUDING
 BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
 DATA, OR PROFITS, OR BUSINESS INTERRUPTION), HOWEVER CAUSED AND ON ANY THEORY
 OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT(INCLUDING
 NEGLIGENCE OR OTHERWISE ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
 EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 -->
<html>
<head>
<meta charset="utf-8">
    <title>WebCLCommandQueue enqueueReadBufferRect conformance test.</title>
    <link rel="stylesheet" href="../../resources/js-test-style.css"/>
    <script src="../../resources/js-test-pre.js"></script>
    <script src="../../resources/cl_utils.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>

try {
    debug("WebCLCommandQueue enqueueReadBufferRect test case.\n");
    debug("( WebCLBuffer buffer, CLboolean blockingRead, CLuint[3] bufferOrigin, CLuint[3] hostOrigin, CLuint[3] region, CLuint bufferRowPitch, CLuint bufferSlicePitch, CLuint hostRowPitch, CLuint hostSlicePitch, ArrayBuffer hostPtr, optional WebCLEvent[]? eventWaitList, optional WebCLEvent? event) ");

    var WIDTH_STEP = Float32Array.BYTES_PER_ELEMENT * 320 * 4;
    var BUFFER_SIZE = Float32Array.BYTES_PER_ELEMENT * 320 * 240 * 4;
    var bufferRowPitch = WIDTH_STEP;
    var bufferSlicePitch = WIDTH_STEP;
    var hostRowPitch = WIDTH_STEP;
    var hostSlicePitch = WIDTH_STEP;
    var wtu = WebCLTestUtils;
    var webCLContext = wtu.createContext();
    var webCLCommandQueue = wtu.createCommandQueue(webCLContext);
    var webCLBufferOrigin = new Int32Array([12, 25, 0]);
    var hostOrigin = new Int32Array([12, 25, 0]);
    var region = new Int32Array([10, 20, 1]);
    var arrBuffer = new Float32Array(BUFFER_SIZE);

    // Testing with buffer as MEM_READ_WRITE memory object.
    var webCLBufferReadWrite = webCLContext.createBuffer(webcl.MEM_READ_WRITE, BUFFER_SIZE);
    shouldBeTrue("wtu.enqueueReadBufferRect(webCLCommandQueue, webCLBufferReadWrite, true, webCLBufferOrigin, hostOrigin, region, bufferRowPitch, bufferSlicePitch, hostRowPitch, hostSlicePitch, arrBuffer);");
    shouldBeTrue("wtu.enqueueReadBufferRect(webCLCommandQueue, webCLBufferReadWrite, false, webCLBufferOrigin, hostOrigin, region, bufferRowPitch, bufferSlicePitch, hostRowPitch, hostSlicePitch, arrBuffer);");

    // Testing with buffer as MEM_READ_ONLY memory object.
    var webCLBufferRead = webCLContext.createBuffer(webcl.MEM_READ_ONLY, BUFFER_SIZE);
    shouldBeTrue("wtu.enqueueReadBufferRect(webCLCommandQueue, webCLBufferRead, true, webCLBufferOrigin, hostOrigin, region, bufferRowPitch, bufferSlicePitch, hostRowPitch, hostSlicePitch, arrBuffer);");
    shouldBeTrue("wtu.enqueueReadBufferRect(webCLCommandQueue, webCLBufferRead, true, webCLBufferOrigin, hostOrigin, region, bufferRowPitch, bufferSlicePitch, hostRowPitch, hostSlicePitch, arrBuffer);");

    // Testing with buffer as MEM_WRITE_ONLY memory object.
    var webCLBufferWrite = webCLContext.createBuffer(webcl.MEM_WRITE_ONLY, BUFFER_SIZE);
    shouldThrowExceptionName("wtu.enqueueReadBufferRect(webCLCommandQueue, webCLBufferWrite, true, webCLBufferOrigin, hostOrigin, region, bufferRowPitch, bufferSlicePitch, hostRowPitch, hostSlicePitch, arrBuffer);", "INVALID_VALUE");
    shouldThrowExceptionName("wtu.enqueueReadBufferRect(webCLCommandQueue, webCLBufferWrite, true, webCLBufferOrigin, hostOrigin, region, bufferRowPitch, bufferSlicePitch, hostRowPitch, hostSlicePitch, arrBuffer);", "INVALID_VALUE");

    var webCLBuffer= webCLContext.createBuffer(webcl.MEM_READ_WRITE, BUFFER_SIZE);

    // Testing for values of Buffer Origin.
    var OriginInvalid = new Int32Array([-12, -25, 0]);
    shouldThrowExceptionName("wtu.enqueueReadBufferRect(webCLCommandQueue, webCLBuffer, true, OriginInvalid, hostOrigin, region, bufferRowPitch, bufferSlicePitch, hostRowPitch, hostSlicePitch, arrBuffer);", "INVALID_VALUE");

    // Testing for values of rectangular region being written.
    var regionInvalid = new Int32Array([-10, -20, 1]);
    shouldThrowExceptionName("wtu.enqueueReadBufferRect(webCLCommandQueue, webCLBuffer, true, webCLBufferOrigin, regionInvalid, bufferRowPitch, bufferSlicePitch, hostRowPitch, hostSlicePitch, arrBuffer);", "INVALID_VALUE");

    // FIXME :: Add test cases for invalid values of Buffer row pitch, Buffer slice pitch, Host row pitch and Host source pitch.

    // Testing invalid objects as params.
    shouldThrowExceptionName("wtu.enqueueReadBufferRect(webCLCommandQueue, invalid_object, true, webCLBufferOrigin, region, bufferRowPitch, bufferSlicePitch, hostRowPitch, hostSlicePitch, arrBuffer);", "INVALID_MEM_OBJECT");
    shouldThrowExceptionName("wtu.enqueueReadBufferRect(webCLCommandQueue, webCLBuffer, invalid_object, webCLOrigin, region, bufferRowPitch, bufferSlicePitch, hostRowPitch, hostSlicePitch, arrBuffer);", "INVALID_VALUE");
    shouldThrowExceptionName("wtu.enqueueReadBufferRect(webCLCommandQueue, webCLBuffer, true, invalid_object, region, bufferRowPitch, bufferSlicePitch, hostRowPitch, hostSlicePitch, arrBuffer);", "INVALID_VALUE");
    shouldThrowExceptionName("wtu.enqueueReadBufferRect(webCLCommandQueue, webCLBuffer, true, webCLBufferOrigin, invalid_object, bufferRowPitch, bufferSlicePitch, hostRowPitch, hostSlicePitch, arrBuffer);", "INVALID_VALUE");
    shouldThrowExceptionName("wtu.enqueueReadBufferRect(webCLCommandQueue, webCLBuffer, true, webCLBufferOrigin, region, invalid_object, bufferSlicePitch, hostRowPitch, hostSlicePitch, arrBuffer);", "INVALID_VALUE");
    shouldThrowExceptionName("wtu.enqueueReadBufferRect(webCLCommandQueue, webCLBuffer, true, webCLBufferOrigin, region, bufferRowPitch, invalid_object, hostRowPitch, hostSlicePitch, arrBuffer);", "INVALID_VALUE");
    shouldThrowExceptionName("wtu.enqueueReadBufferRect(webCLCommandQueue, webCLBuffer, true, webCLBufferOrigin, region, bufferRowPitch, bufferSlicePitch, invalid_object, hostSlicePitch, arrBuffer);", "INVALID_VALUE");
    shouldThrowExceptionName("wtu.enqueueReadBufferRect(webCLCommandQueue, webCLBuffer, true, webCLBufferOrigin, region, invalid_object, bufferRowPitch, bufferSlicePitch, hostRowPitch, invalid_object, arrBuffer);", "INVALID_VALUE");
    shouldThrowExceptionName("wtu.enqueueReadBufferRect(webCLCommandQueue, webCLBuffer, true, webCLBufferOrigin, region, invalid_object, bufferRowPitch, bufferSlicePitch, hostRowPitch, hostSlicePitch, invalid_object);", "INVALID_VALUE");

    // FIXME :: Need to add test cases passing event parameters.

    // Testing for exception "INVALID_CONTEXT".
    var webCLContext2 = wtu.createContext();
    var webCLCommandqueue2 = wtu.createCommandQueue(webCLContext2);
    shouldThrowExceptionName("wtu.enqueueReadBufferRect(webCLCommandqueue2, webCLBuffer, true, webCLBufferOrigin, region, bufferRowPitch, bufferSlicePitch, hostRowPitch, hostSlicePitch, arrBuffer);", "INVALID_CONTEXT");

} catch (e) {
    testFailed("WebCLCommandQueue :: enqueueReadBufferRect throws an exception : " + e.message);
}
notifyFinishedToHarness();
successfullyParsed = true;
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
